+++ plugins/system/cwgears/helpers/loadcount.php
@@ -1,16 +1,34 @@
 <?php
 
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+defined('_JEXEC') or die('Restricted access');
+
+/**
+ * @package             Joomla
+ * @subpackage          CoalaWeb Gears
+ * @author              Steven Palmer
+ * @author url          http://coalaweb.com
+ * @author email        support@coalaweb.com
+ * @license             GNU/GPL, see /files/en-GB.license.txt
+ * @copyright           Copyright (c) 2015 Steven Palmer All rights reserved.
+ *
+ * CoalaWeb Gears is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 /**
- * Description of loadcount
- *
- * @author steve
+ *  component helper.
  */
+
 class CwGearsHelperLoadcount {
 
     /**
@@ -40,6 +58,10 @@ class CwGearsHelperLoadcount {
     public static function setUikitCount($url) {
         //Start our database queries
         $db = JFactory::getDbo();
+        //Current date time
+        $siteOffset = JFactory::getApplication()->getCfg('offset');
+        $dtnow = JFactory::getDate('now', $siteOffset);
+        $now = $dtnow->toUnix(true);
 
         // Check if page already exists then update else insert
         $query = $db->getQuery(true);
@@ -54,7 +76,8 @@ class CwGearsHelperLoadcount {
                 $query = $db->getQuery(true);
 
                 $fields = array(
-                    $db->quoteName('uikit') . ' = 1'
+                    $db->quoteName('uikit') . ' = '. $db->quote('1'),
+                    $db->quoteName('time') . ' = ' . $db->quote($now)
                 );
 
                 $conditions = array(
@@ -74,16 +97,17 @@ class CwGearsHelperLoadcount {
                 $query = $db->getQuery(true);
                 $query->insert($db->quoteName('#__cwgears'));
                 $query->columns(
-                        'url, uikit ');
+                        'url, uikit, time');
                 $query->values(
                         $db->quote($url) . ','
-                        . $db->quote('1'));
+                        . $db->quote('1') . ','
+                        . $db->quote($now));
 
                 $db->setQuery($query);
                 $db->execute();
             } catch (Exception $e) {
-                JLog::addLogger(array('text_file' => 'plg_cwgears.error.php'));
-                JLog::add($e->getMessage());
+//                JLog::addLogger(array('text_file' => 'plg_cwgears.error.php'));
+//                JLog::add($e->getMessage());
             }
         }
 
@@ -93,6 +117,10 @@ class CwGearsHelperLoadcount {
     public static function setUikitPlusCount($url) {
         //Start our database queries
         $db = JFactory::getDbo();
+        //Current date time
+        $siteOffset = JFactory::getApplication()->getCfg('offset');
+        $dtnow = JFactory::getDate('now', $siteOffset);
+        $now = $dtnow->toUnix(true);
 
         // Check if page already exists then update else insert
         $query = $db->getQuery(true);
@@ -107,7 +135,8 @@ class CwGearsHelperLoadcount {
                 $query = $db->getQuery(true);
 
                 $fields = array(
-                    $db->quoteName('uikit_plus') . ' = 1'
+                    $db->quoteName('uikit_plus') . ' = ' .$db->quote('1'),
+                    $db->quoteName('time') . ' = ' . $db->quote($now)
                 );
 
                 $conditions = array(
@@ -119,24 +148,25 @@ class CwGearsHelperLoadcount {
                 $db->setQuery($query);
                 $db->execute();
             } catch (Exception $e) {
-                JLog::addLogger(array('text_file' => 'plg_cwgears.error.php'));
-                JLog::add($e->getMessage());
+//                JLog::addLogger(array('text_file' => 'plg_cwgears.error.php'));
+//                JLog::add($e->getMessage());
             }
         } else {//we are inserting
             try {
                 $query = $db->getQuery(true);
                 $query->insert($db->quoteName('#__cwgears'));
                 $query->columns(
-                        'url, uikit_plus ');
+                        'url, uikit_plus, time');
                 $query->values(
                         $db->quote($url) . ','
-                        . $db->quote('1'));
+                        . $db->quote('1') . ','
+                        . $db->quote($now));
 
                 $db->setQuery($query);
                 $db->execute();
             } catch (Exception $e) {
-                JLog::addLogger(array('text_file' => 'plg_cwgears.error.php'));
-                JLog::add($e->getMessage());
+//                JLog::addLogger(array('text_file' => 'plg_cwgears.error.php'));
+//                JLog::add($e->getMessage());
             }
         }
 
@@ -149,6 +179,10 @@ class CwGearsHelperLoadcount {
     public static function setFacebookJsCount($url) {
         //Start our database queries
         $db = JFactory::getDbo();
+        //Current date time
+        $siteOffset = JFactory::getApplication()->getCfg('offset');
+        $dtnow = JFactory::getDate('now', $siteOffset);
+        $now = $dtnow->toUnix(true);
 
         // Check if page already exists then update else insert
         $query = $db->getQuery(true);
@@ -163,7 +197,8 @@ class CwGearsHelperLoadcount {
                 $query = $db->getQuery(true);
 
                 $fields = array(
-                    $db->quoteName('facebook_js') . ' = 1'
+                    $db->quoteName('facebook_js') . ' = ' .$db->quote('1'),
+                    $db->quoteName('time') . ' = ' . $db->quote($now)
                 );
 
                 $conditions = array(
@@ -175,24 +210,25 @@ class CwGearsHelperLoadcount {
                 $db->setQuery($query);
                 $db->execute();
             } catch (Exception $e) {
-                JLog::addLogger(array('text_file' => 'plg_cwgears.error.php'));
-                JLog::add($e->getMessage());
+//                JLog::addLogger(array('text_file' => 'plg_cwgears.error.php'));
+//                JLog::add($e->getMessage());
             }
         } else {//we are inserting
             try {
                 $query = $db->getQuery(true);
                 $query->insert($db->quoteName('#__cwgears'));
                 $query->columns(
-                        'url, facebook_js ');
+                        'url, facebook_js, time ');
                 $query->values(
                         $db->quote($url) . ','
-                        . $db->quote('1'));
+                        . $db->quote('1') . ','
+                        . $db->quote($now));
 
                 $db->setQuery($query);
                 $db->execute();
             } catch (Exception $e) {
-                JLog::addLogger(array('text_file' => 'plg_cwgears.error.php'));
-                JLog::add($e->getMessage());
+//                JLog::addLogger(array('text_file' => 'plg_cwgears.error.php'));
+//                JLog::add($e->getMessage());
             }
         }