+++ plugins/system/cwgears/sql/install/mysql/uninstall.mysql.utf8.sql
@@ -1,1 +1,2 @@
 DROP TABLE IF EXISTS `#__cwgears`;
+DROP TABLE IF EXISTS `#__cwgears_schedule`;