+++ plugins/system/cwgears/sql/updates/mysql/0.1.5.sql
@@ -1,1 +1,8 @@
-# Dummy SQL file to set schema version to 0.1.6 so next update will work
+CREATE TABLE IF NOT EXISTS `#__cwgears` (
+  `id` int(11) unsigned NOT NULL AUTO_INCREMENT COMMENT 'Primary Key',
+  `url` varchar(255) NOT NULL,
+  `facebook_js` int(11) NOT NULL,
+  `uikit` int(11) NOT NULL,
+  `uikit_plus` int(11) NOT NULL,
+  PRIMARY KEY  (`id`)
+) ENGINE=MyISAM DEFAULT CHARSET=utf8;